+++ js/tokens.ts
@@ -47,6 +47,16 @@ export const DEFAULT_WIDTHS = {
   node: 1,
 };
 
+export const DEFAULT_CURVATURES = {
+  link: 0,
+  selected: 0,
+};
+
+export const DEFAULT_LINE_DASHES = {
+  link: [],
+  selected: [],
+};
+
 export const WIDGET_DEFAULTS = {
   _model_module: NAME,
   _model_module_version: VERSION,
@@ -69,6 +79,8 @@ export interface IBehave {
   extraColumns?: IExtraColumns;
   // link
   getLinkColor?(options: ILinkBehaveOptions): string | null;
+  getLinkCurvature?(options: ILinkBehaveOptions): number | null;
+  getLinkLineDash?(options: ILinkBehaveOptions): number[] | null;
   getLinkWidth?(options: ILinkBehaveOptions): number | null;
   getLinkLabel?(options: ILinkBehaveOptions): string | null;
   getLinkDirectionalArrowColor?(options: ILinkBehaveOptions): string | null;
@@ -93,6 +105,8 @@ export interface IBehave {
 export const ALL_LINK_METHODS = [
   'getLinkLabel',
   'getLinkColor',
+  'getLinkCurvature',
+  'getLinkLineDash',
   'getLinkWidth',
   'getLinkDirectionalArrowColor',
   'getLinkDirectionalArrowLength',
@@ -220,6 +234,7 @@ export const FALSEY = Object.freeze([
 ]);
 
 export enum ECoerce {
+  array = 'array',
   boolish = 'boolean',
   numeric = 'number',
 }