+++ src/ipyforcegraph/__init__.py
@@ -1,4 +1,4 @@
-"""Main entrypoint for ``ipyforcegraph``."""
+"""Main module metadata for ``ipyforcegraph``."""
 
 # Copyright (c) 2023 ipyforcegraph contributors.
 # Distributed under the terms of the Modified BSD License.
@@ -9,6 +9,7 @@ from .constants import EXTENSION_NAME, __version__
 
 
 def _jupyter_labextension_paths() -> List[Dict[str, str]]:
+    """Magic method for finding JupyterLab extension static assets."""
     from .js import __prefix__
 
     return [dict(src=str(__prefix__), dest=EXTENSION_NAME)]