+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## `0.3.0` (unreleased)
+## `0.3.1`
+
+### `ipyforcegraph 0.3.1`
+
+- fixes compatibility with `jupyterlab_widgets 3.0.6`
+
+### `@jupyrdf/jupyter-forcegraph 0.3.1`
+
+## `0.3.0`
 
 ### `ipyforcegraph 0.3.0`